Refactor/further deduplications (#311)
* Attempting further deduplications.

* Reverting change to the partial updates.

* Moving functions for PR clarity.

* Further refactors.

* Deduplicating further by parameterising on partial update callable.

* Fixing failing tests.

---------
