+++ deployments.py
@@ -14,6 +14,7 @@ from prefect.client.schemas.schedules import CronSchedule
 from prefect.deployments.runner import DeploymentImage
 from prefect.flows import Flow
 
+from flows.boundary import run_partial_updates_of_concepts_for_batch
 from flows.count_family_document_concepts import (
     count_family_document_concepts,
     load_update_document_concepts_counts,
@@ -22,8 +23,7 @@ from flows.data_backup import data_backup
 from flows.deploy_static_sites import deploy_static_sites
 from flows.index import (
     index_labelled_passages_from_s3_to_vespa,
-    run_partial_updates_of_concepts_for_batch,
-    run_partial_updates_of_concepts_for_document_passages,
+    run_partial_updates_of_concepts_for_document_passages__update,
 )
 from flows.inference import (
     classifier_inference,
@@ -106,7 +106,7 @@ create_deployment(
 # Index
 
 create_deployment(
-    flow=run_partial_updates_of_concepts_for_document_passages,
+    flow=run_partial_updates_of_concepts_for_document_passages__update,
     description="Co-ordinate updating inference results for concepts in Vespa",
 )
