+++ flows/index.py
@@ -1,33 +1,28 @@
 import asyncio
 import json
 import os
-import re
 from collections import Counter
 from dataclasses import dataclass
 from pathlib import Path
 from typing import Any
 
-import vespa.querybuilder as qb
 from cpr_sdk.models.search import Concept as VespaConcept
 from cpr_sdk.models.search import Passage as VespaPassage
 from cpr_sdk.search_adaptors import VespaSearchAdapter
 from prefect import flow
-from prefect.deployments.deployments import run_deployment
-from prefect.logging import get_logger, get_run_logger
-from pydantic import BaseModel
-from vespa.io import VespaQueryResponse, VespaResponse
+from prefect.logging import get_run_logger
 
 from flows.boundary import (
+    DEFAULT_DOCUMENTS_BATCH_SIZE,
+    DEFAULT_INDEXING_TASK_BATCH_SIZE,
+    ConceptModel,
     DocumentImporter,
-    DocumentImportId,
-    DocumentStem,
     TextBlockId,
-    VespaHitId,
     convert_labelled_passage_to_concepts,
-    get_data_id_from_vespa_hit_id,
     get_vespa_search_adapter,
+    index_by_s3,
     load_labelled_passages_by_uri,
-    s3_obj_generator,
+    partial_update_text_block,
     s3_object_write_text,
     s3_paths_or_s3_prefixes,
 )
@@ -40,19 +35,12 @@ from flows.utils import (
 from scripts.cloud import (
     AwsEnv,
     ClassifierSpec,
-    function_to_flow_name,
-    generate_deployment_name,
     get_prefect_job_variable,
 )
 from scripts.update_classifier_spec import parse_spec_file
-from src.exceptions import PartialUpdateError, QueryError
 from src.identifiers import WikibaseID
 
-DEFAULT_DOCUMENTS_BATCH_SIZE = 500
-DEFAULT_INDEXING_TASK_BATCH_SIZE = 20
-HTTP_OK = 200
 CONCEPTS_COUNTS_PREFIX_DEFAULT: str = "concepts_counts"
-CONCEPT_COUNT_SEPARATOR: str = ":"
 
 
 @dataclass()
@@ -93,142 +81,7 @@ class Config:
         return config
 
 
-class ConceptModel(BaseModel):
-    """
-    A concept and the model used to identify it.
-
-    This class represents a pairing of a Wikibase concept ID with the name of the model
-    used to identify that concept in text. It is hashable to allow use in sets and as
-    dictionary keys.
-
-    Attributes:
-        wikibase_id: The Wikibase ID of the concept
-        model_name: The name of the model used to identify the concept
-
-    Example:
-        >>> model = ConceptModel(wikibase_id=WikibaseID("Q123"),
-        ...                     model_name='KeywordClassifier("professional services sector")')
-        >>> str(model)
-        'Q123:KeywordClassifier("professional services sector")'
-    """
-
-    wikibase_id: WikibaseID
-    model_name: str
-
-    def __str__(self) -> str:
-        """
-        Convert the ConceptModel to a string in the format 'WIKIBASE_ID:MODEL_NAME'.
-
-        Returns:
-            str: String representation in the format 'WIKIBASE_ID:MODEL_NAME'
-        """
-        return f"{self.wikibase_id}{CONCEPT_COUNT_SEPARATOR}{self.concept_name}"
-
-    def __hash__(self) -> int:
-        """
-        Generate a hash value for the ConceptModel.
-
-        The hash is based on both the wikibase_id and model_name to ensure
-        uniqueness when used in sets or as dictionary keys.
-
-        Returns:
-            int: Hash value of the ConceptModel
-        """
-        return hash((self.wikibase_id, self.model_name))
-
-    def __eq__(self, other: object) -> bool:
-        """
-        Compare this ConceptModel with another for equality.
-
-        Two ConceptModels are equal if they have the same wikibase_id and model_name.
-
-        Args:
-            other: The object to compare with
-
-        Returns:
-            bool: True if the objects are equal, False otherwise
-        """
-        if not isinstance(other, ConceptModel):
-            return NotImplemented
-        return (
-            self.wikibase_id == other.wikibase_id
-            and self.model_name == other.model_name
-        )
-
-    @property
-    def concept_name(self) -> str:
-        """
-        Extract the concept name from the model name.
-
-        Examples:
-            >>> model = ConceptModel(wikibase_id=WikibaseID("Q123"),
-            ...                     model_name='KeywordClassifier("professional services sector")')
-            >>> model.concept_name
-            'professional services sector'
-            >>> model2 = ConceptModel(wikibase_id=WikibaseID("Q456"),
-            ...                      model_name='LLMClassifier("agriculture")')
-            >>> model2.concept_name
-            'agriculture'
-            >>> model3 = ConceptModel(wikibase_id=WikibaseID("Q789"),
-            ...                      model_name='InvalidModelName')
-            >>> model3.concept_name  # doctest: +IGNORE_EXCEPTION_DETAIL
-            Traceback (most recent call last):
-                ...
-            ValueError: Could not extract concept name from model name 'InvalidModelName'
-        """
-        match = re.search(r'"([^"]+)"', self.model_name)
-        if match:
-            return match.group(1)
-
-        raise ValueError(
-            f"Could not extract concept name from model name '{self.model_name}'"
-        )
-
-
-def get_document_passage_from_vespa(
-    text_block_id: str,
-    document_import_id: DocumentImportId,
-    vespa_search_adapter: VespaSearchAdapter,
-) -> tuple[VespaHitId, VespaPassage]:
-    """Retrieve a passage for a document in Vespa."""
-    logger = get_logger()
-
-    logger.info(
-        f"Getting document passage from Vespa: {document_import_id}, text block: {text_block_id}"
-    )
-
-    condition = qb.QueryField("family_document_ref").contains(
-        f"id:doc_search:family_document::{document_import_id}"
-    ) & qb.QueryField("text_block_id").contains(text_block_id)
-
-    yql = qb.select("*").from_("document_passage").where(condition)
-
-    vespa_query_response: VespaQueryResponse = vespa_search_adapter.client.query(
-        yql=yql
-    )
-
-    if not vespa_query_response.is_successful():
-        raise QueryError(vespa_query_response.get_status_code())
-    if len(vespa_query_response.hits) != 1:
-        raise ValueError(
-            f"Expected 1 document passage for text block `{text_block_id}`, got {len(vespa_query_response.hits)}"
-        )
-
-    logger.info(
-        (
-            f"Vespa search response for document: {document_import_id} "
-            f"with {len(vespa_query_response.hits)} hits"
-        )
-    )
-
-    hit = vespa_query_response.hits[0]
-    passage_id = hit["id"]
-    passage = VespaPassage.model_validate(hit["fields"])
-
-    return passage_id, passage
-
-
-def get_updated_passage_concepts(
+def update_concepts_on_existing_vespa_concepts(
     passage: VespaPassage,
     concepts: list[VespaConcept],
 ) -> list[dict[str, Any]]:
@@ -258,8 +111,12 @@ def get_updated_passage_concepts(
     return [concept_.model_dump(mode="json") for concept_ in updated_concepts]
 
 
+# FIXME: From what I can tell this function should be identical to the related function
+# deindex.py. In deindex.py we need to remove_translated_suffix from the
+# document_import_id and in this function I'm assuming we should add the updates to how
+# we calculate concepts counts from results.
 @flow
-async def run_partial_updates_of_concepts_for_document_passages(
+async def run_partial_updates_of_concepts_for_document_passages__update(
     document_importer: DocumentImporter,
     cache_bucket: str,
     concepts_counts_prefix: str,
@@ -316,6 +173,7 @@ async def run_partial_updates_of_concepts_for_document_passages(
                     concepts=concepts,
                     document_import_id=document_import_id,
                     vespa_search_adapter=vespa_search_adapter,
+                    update_function=update_concepts_on_existing_vespa_concepts,
                 )
                 for text_block_id, concepts in batch
             ]
@@ -380,183 +238,6 @@ async def run_partial_updates_of_concepts_for_document_passages(
         return concepts_counts
 
 
-async def partial_update_text_block(
-    text_block_id: TextBlockId,
-    concepts: list[VespaConcept],  # A possibly empty list
-    document_import_id: DocumentImportId,
-    vespa_search_adapter: VespaSearchAdapter,
-):
-    """Partial update a singular text block and its concepts."""
-    document_passage_id, document_passage = get_document_passage_from_vespa(
-        text_block_id, document_import_id, vespa_search_adapter
-    )
-
-    data_id = get_data_id_from_vespa_hit_id(document_passage_id)
-
-    serialised_concepts = get_updated_passage_concepts(
-        document_passage,
-        concepts,
-    )
-
-    response: VespaResponse = vespa_search_adapter.client.update_data(  # pyright: ignore[reportOptionalMemberAccess]
-        schema="document_passage",
-        namespace="doc_search",
-        data_id=data_id,
-        fields={"concepts": serialised_concepts},
-    )
-
-    if (status_code := response.get_status_code()) != HTTP_OK:
-        raise PartialUpdateError(data_id, status_code)
-
-
-@flow
-async def run_partial_updates_of_concepts_for_batch(
-    documents_batch: list[DocumentImporter],
-    documents_batch_num: int,
-    cache_bucket: str,
-    concepts_counts_prefix: str,
-) -> None:
-    """Run partial updates for concepts in a batch of documents."""
-
-    logger = get_run_logger()
-    logger.info(
-        f"Updating concepts for batch of documents, documents in batch: {len(documents_batch)}."
-    )
-    for i, document_importer in enumerate(documents_batch):
-        try:
-            _ = await run_partial_updates_of_concepts_for_document_passages(
-                document_importer=document_importer,
-                cache_bucket=cache_bucket,
-                concepts_counts_prefix=concepts_counts_prefix,
-            )
-
-            logger.info(f"processed batch documents #{documents_batch_num}")
-
-        except Exception as e:
-            document_stem: DocumentStem = documents_batch[i][0]
-            logger.error(
-                f"failed to process document `{document_stem}`: {e.__str__()}",
-            )
-            continue
-
-
-async def run_partial_updates_of_concepts_for_batch_flow_or_deployment(
-    documents_batch: list[DocumentImporter],
-    documents_batch_num: int,
-    cache_bucket: str,
-    concepts_counts_prefix: str,
-    aws_env: AwsEnv,
-    as_deployment: bool,
-) -> None:
-    """Run partial updates for a batch of documents as a sub-flow or deployment."""
-    logger = get_run_logger()
-    logger.info(
-        "Running partial updates of concepts for batch as sub-flow or deployment: "
-        f"batch length {len(documents_batch)}, as_deployment: {as_deployment}"
-    )
-
-    if as_deployment:
-        flow_name = function_to_flow_name(run_partial_updates_of_concepts_for_batch)
-        deployment_name = generate_deployment_name(flow_name=flow_name, aws_env=aws_env)
-
-        return await run_deployment(
-            name=f"{flow_name}/{deployment_name}",
-            parameters={
-                "documents_batch": documents_batch,
-                "documents_batch_num": documents_batch_num,
-                "cache_bucket": cache_bucket,
-                "concepts_counts_prefix": concepts_counts_prefix,
-            },
-            timeout=3600,
-        )
-
-    return await run_partial_updates_of_concepts_for_batch(
-        documents_batch=documents_batch,
-        documents_batch_num=documents_batch_num,
-        cache_bucket=cache_bucket,
-        concepts_counts_prefix=concepts_counts_prefix,
-    )
-
-
-@flow
-async def index_by_s3(
-    aws_env: AwsEnv,
-    cache_bucket: str,
-    concepts_counts_prefix: str,
-    vespa_search_adapter: VespaSearchAdapter | None,
-    s3_prefixes: list[str] | None = None,
-    s3_paths: list[str] | None = None,
-    batch_size: int = DEFAULT_DOCUMENTS_BATCH_SIZE,
-    indexing_task_batch_size: int = DEFAULT_INDEXING_TASK_BATCH_SIZE,
-    as_deployment: bool = True,
-) -> None:
-    """
-    Asynchronously index concepts from S3 files into Vespa.
-
-    This function retrieves concept documents from files stored in an S3 path and
-    indexes them in a Vespa instance. The name of each file in the specified S3 path is
-    expected to represent the document's import ID.
-
-    When `s3_prefix` is provided, the function will index all files within that S3
-    prefix (directory). When `s3_paths` is provided, the function will index only the
-    files specified in the list of S3 object keys. If both are provided `s3_paths` will
-    be used.
-
-    Assumptions:
-    - The S3 file names represent document import IDs.
-
-    params:
-    - s3_prefix: The S3 prefix (directory) to yield objects from.
-        E.g. "s3://bucket/prefix/"
-    - s3_paths: A list of S3 object keys to yield objects from.
-        E.g. {"s3://bucket/prefix/file1.json", "s3://bucket/prefix/file2.json"}
-    - vespa_search_adapter: An instance of VespaSearchAdapter.
-        E.g. VespaSearchAdapter(
-            instance_url="https://vespa-instance-url.com",
-            cert_directory="certs/"
-        )
-    """
-    logger = get_run_logger()
-
-    cm, vespa_search_adapter = get_vespa_search_adapter(vespa_search_adapter)
-
-    with cm:
-        logger.info("Getting S3 object generator")
-        documents_generator = s3_obj_generator(s3_prefixes, s3_paths)
-        documents_batches = iterate_batch(documents_generator, batch_size=batch_size)
-        indexing_task_batches = iterate_batch(
-            data=documents_batches, batch_size=indexing_task_batch_size
-        )
-
-        for i, indexing_task_batch in enumerate(indexing_task_batches, start=1):
-            logger.info(f"Processing indexing task batch #{i}")
-
-            indexing_tasks = [
-                run_partial_updates_of_concepts_for_batch_flow_or_deployment(
-                    documents_batch=documents_batch,
-                    documents_batch_num=documents_batch_num,
-                    cache_bucket=cache_bucket,
-                    concepts_counts_prefix=concepts_counts_prefix,
-                    aws_env=aws_env,
-                    as_deployment=as_deployment,
-                )
-                for documents_batch_num, documents_batch in enumerate(
-                    indexing_task_batch, start=1
-                )
-            ]
-
-            logger.info(f"Gathering indexing tasks for batch #{i}")
-            results = await asyncio.gather(*indexing_tasks, return_exceptions=True)
-            logger.info(f"Gathered indexing tasks for batch #{i}")
-
-            for result in results:
-                if isinstance(result, Exception):
-                    logger.error(
-                        f"failed to process document batch in indexing task batch #{i}: {str(result)}",
-                    )
-                    continue
-
-
 @flow(
     on_failure=[SlackNotify.message],
     on_crashed=[SlackNotify.message],
@@ -604,6 +285,7 @@ async def index_labelled_passages_from_s3_to_vespa(
     logger.info(f"s3_prefixes: {s3_accessor.prefixes}, s3_paths: {s3_accessor.paths}")
 
     await index_by_s3(
+        partial_update_flow=run_partial_updates_of_concepts_for_document_passages__update,
         aws_env=config.aws_env,
         vespa_search_adapter=config.vespa_search_adapter,
         s3_prefixes=s3_accessor.prefixes,
